Fix Stripe Terminal extraction by adding terminal.rs to Bazel build targets
Resolves #29330.

The `terminal.rs` module was previously extracted, but it was missing from Bazel's build graphs, leading to `file not found for module terminal` errors. Added the module to `srcs` and `exports_files` in `src/server/BUILD.bazel`, `src/server/integrations/BUILD.bazel`, and `src/server/integrations/stripe/BUILD.bazel`.

Also cleaned up unused references in `bazel_lib.rs` and properly used `super::client::StripeClient` instead of `crate::integrations::stripe::client::StripeClient` in `terminal.rs` for Bazel compatibility.
