+++ src/server/BUILD.bazel
@@ -147,6 +147,7 @@ SERVER_RS_SRCS = [
     "//src/server/integrations/stripe:mod.rs",
     "//src/server/integrations/stripe:payout_batcher.rs",
     "//src/server/integrations/stripe:routing.rs",
+    "//src/server/integrations/stripe:terminal.rs",
     "//src/server/interop:mod.rs",
     "//src/server/interop:protocol.rs",
     "//src/server/orchestration:chaos_test.rs",
