+++ src/server/integrations/stripe/BUILD.bazel
@@ -8,6 +8,7 @@ exports_files(
         "mod.rs",
         "payout_batcher.rs",
         "routing.rs",
+        "terminal.rs",
     ],
     visibility = ["//visibility:public"],
 )
@@ -20,6 +21,7 @@ rust_library(
         "mod.rs",
         "payout_batcher.rs",
         "routing.rs",
+        "terminal.rs",
     ],
     crate_root = "bazel_lib.rs",
     edition = "2024",
