+++ src/server/integrations/stripe/mod.rs
@@ -1,3 +1,6 @@
 pub mod client;
 pub mod routing;
 pub mod payout_batcher;
+
+
+pub mod terminal;
