+++ src/server/integrations/stripe/terminal.rs
@@ -1,14 +1,30 @@
-use crate::integrations::stripe::client::StripeClient;
+use super::client::StripeClient;
 
 impl StripeClient {
-    pub async fn create_terminal_connection_token(&self, tenant_id: &str) -> Result<String, String> {
+    pub async fn create_terminal_connection_token(&self, _tenant_id: &str) -> Result<String, String> {
+        let api_key = self.require_api_key()?;
+        let res = reqwest::Client::new()
+            .post(format!("{}/v1/terminal/connection_tokens", Self::api_base()))
+            .basic_auth(api_key, Some(""))
+            .form(&std::collections::HashMap::<String, String>::new())
+            .send()
+            .await
+            .map_err(|e| format!("Stripe Terminal connection token request failed: {}", e))?;
 
-        // In a real implementation, this would make an HTTP POST to Stripe's /v1/terminal/connection_tokens
-        // endpoint. Since we're mocking external APIs, we return a mock token string here.
-        // We simulate the token being tightly scoped to the tenant for multi-tenant isolation.
-        let mock_token = format!("tss_mock_token_for_{}", tenant_id);
+        if !res.status().is_success() {
+            let status = res.status();
+            let text = res.text().await.unwrap_or_default();
+            return Err(format!("Stripe Terminal API error ({}): {}", status, text));
+        }
 
-        Ok(mock_token)
+        let json: serde_json::Value = res
+            .json()
+            .await
+            .map_err(|e| format!("Failed to parse Stripe Terminal token response: {}", e))?;
+        json["secret"]
+            .as_str()
+            .map(|secret| secret.to_string())
+            .ok_or_else(|| "Missing secret in Stripe Terminal token response".to_string())
     }
 
     pub async fn create_terminal_payment_intent(
@@ -16,17 +32,53 @@ impl StripeClient {
         tenant_id: &str,
         amount_cents: i64,
         currency: &str,
-        _product_id: Option<&str>,
-        _quantity: Option<i32>,
-        _order_id: Option<&str>,
+        product_id: Option<&str>,
+        quantity: Option<i32>,
+        order_id: Option<&str>,
     ) -> Result<String, String> {
+        let api_key = self.require_api_key()?;
+        if amount_cents <= 0 {
+            return Err("amount_cents must be positive".to_string());
+        }
+        if currency.trim().is_empty() {
+            return Err("currency is required".to_string());
+        }
 
-        // In a real implementation, this would make an HTTP POST to Stripe's /v1/payment_intents
-        // endpoint with specific parameters like payment_method_types=["card_present"] and capture_method="manual".
-        // Since we're mocking external APIs, we return a mock intent ID here.
-        let mock_intent = format!("pi_mock_intent_for_{}_{}_{}", tenant_id, amount_cents, currency);
+        let mut form = std::collections::HashMap::new();
+        form.insert("amount".to_string(), amount_cents.to_string());
+        form.insert("currency".to_string(), currency.to_string());
+        form.insert("payment_method_types[]".to_string(), "card_present".to_string());
+        form.insert("capture_method".to_string(), "manual".to_string());
+        form.insert("metadata[tenant_id]".to_string(), tenant_id.to_string());
+        form.insert("metadata[source]".to_string(), "in_person".to_string());
 
-        Ok(mock_intent)
+        if let Some(pid) = product_id {
+            form.insert("metadata[product_id]".to_string(), pid.to_string());
+        }
+        if let Some(qty) = quantity {
+            form.insert("metadata[quantity]".to_string(), qty.to_string());
+        }
+        if let Some(oid) = order_id {
+            form.insert("metadata[order_id]".to_string(), oid.to_string());
+        }
+
+        let res = reqwest::Client::new().post(format!("{}/v1/payment_intents", Self::api_base()))
+            .basic_auth(api_key, Some(""))
+            .form(&form)
+            .send()
+            .await
+            .map_err(|e| format!("Stripe API request failed: {}", e))?;
+
+        if !res.status().is_success() {
+            let status = res.status();
+            let text = res.text().await.unwrap_or_default();
+            return Err(format!("Stripe API error ({}): {}", status, text));
+        }
+
+        let json: serde_json::Value = res.json().await.map_err(|e| format!("Failed to parse response: {}", e))?;
+        let secret = json["client_secret"].as_str().ok_or_else(|| "Missing client_secret in response".to_string())?;
+
+        Ok(secret.to_string())
     }
 }
 
@@ -35,20 +87,10 @@ mod tests {
     use super::*;
 
     #[tokio::test]
-    async fn test_create_terminal_connection_token() {
-        let client = StripeClient::new("sk_test_123".to_string());
+    async fn test_terminal_connection_token_requires_configured_key() {
+        let client = StripeClient::new("".to_string());
         let result = client.create_terminal_connection_token("test_tenant").await;
-        assert!(result.is_ok());
-        let token = result.unwrap();
-        assert_eq!(token, "tss_mock_token_for_test_tenant");
-    }
-
-    #[tokio::test]
-    async fn test_create_terminal_payment_intent() {
-        let client = StripeClient::new("sk_test_123".to_string());
-        let result = client.create_terminal_payment_intent("test_tenant", 1500, "usd", None, None, None).await;
-        assert!(result.is_ok());
-        let intent = result.unwrap();
-        assert_eq!(intent, "pi_mock_intent_for_test_tenant_1500_usd");
+        let err = result.expect_err("Terminal tokens must not be mocked when Stripe credentials are missing");
+        assert!(err.contains("Stripe API key"));
     }
 }
